requirements.txt
websocket-client>=1.7.0
python-dotenv>=1.0.0
pyaudio>=0.2.14
numpy>=1.24.0
scipy>=1.4.1
requests>=2.31.0

# Computer Vision (para detección de objetos)
opencv-contrib-python>=4.12.0
ultralytics>=8.0.196
# mediapipe>=0.10.14  # Opcional - solo para face tracking (requiere Python <3.12)
